feat: add pop-up for create columns
